services/rc_input/requirements.txt
lgpio==0.2.2.0
paho-mqtt==1.6.1
